requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
motor>=3.3.2
solana>=0.30.2
pymongo>=4.6.1
aiohttp>=3.9.1
web3>=6.11.3  # For future multi-chain support
base58>=2.1.1  # For Solana address validation
solders>=0.19.0